Move r generator to sub folder
