+++ src/index.ts
@@ -62,7 +62,7 @@ export {
 } from "./generators/powershell.js";
 export { toPython, toPythonWarn } from "./generators/python/python.js";
 export { toPythonHttp, toPythonHttpWarn } from "./generators/python/http.js";
-export { toR, toRWarn } from "./generators/r.js";
+export { toR, toRWarn } from "./generators/r/r.js";
 export { toRuby, toRubyWarn } from "./generators/ruby.js";
 export { toRust, toRustWarn } from "./generators/rust.js";
 export { toSwift, toSwiftWarn } from "./generators/swift.js";
